fix(lib,python): move libquil.so load to init_libquilc
Originally had it in the Python bindings, but if Python bindings
(e.g. qcs-sdk-rust) depend on this library, they won't necessarily
import the Python code and thus won't hit the path that loads
libquilc.so.

By having it in the library's init, we can ensure that libquilc.so
is always loaded.